Money converted from float to int.

+++ giepp/src/pl/pisz/airlog/giepp/data/CurrentStock.java
@@ -10,19 +10,19 @@ package pl.pisz.airlog.giepp.data;
 public class CurrentStock extends ArchivedStock{
 
 	private String time;
-	private float startPrice;	
-	private float endPrice;
+	private int startPrice;	
+	private int endPrice;
 	private float change;
 	
-	public CurrentStock(String name, String time, float startPrice, float minPrice, float maxPrice, float endPrice, float change) {
+	public CurrentStock(String name, String time, int startPrice, int minPrice, int maxPrice, int endPrice, float change) {
 		super(name,maxPrice,minPrice);
 		this.time = time;
 		this.startPrice = startPrice; 
 		this.endPrice = endPrice; 
 		this.change = change;
 	}
-	public float getStartPrice() { return startPrice; }
-	public float getEndPrice() { return endPrice; }
+	public int getStartPrice() { return startPrice; }
+	public int getEndPrice() { return endPrice; }
 	public float getChange() { return change; }
 	public String getTime() {return time; }
 }
